Added models and moved passport to separate file
