+++ models/userInfo.js
@@ -1,0 +1,8 @@
+const Schema = mongoose.Schema;
+const UserDetail = new Schema({
+    username: String,
+    password: String
+});
+const UserDetails = mongoose.model('userInfo', UserDetail, 'userInfo');
+
+module.exports = UserDetails;
